requirements.txt
# Core ML Libraries
scikit-learn
pandas
numpy

# FastAPI and Server
fastapi
uvicorn[standard]
pydantic

# Supabase
supabase

# Environment Variables
python-dotenv

# Additional utilities
python-multipart
